+++ src/rules/index.ts
@@ -1,7 +1,7 @@
 export * as base from "./base.js";
 export * as environment from "./environment.js";
 export * as importPlugin from "./import.js";
-export * as importSort from "./import-sort.js";
+export * as importSort from "./importSort.js";
 export * as node from "./node.js";
 export * as react from "./react.js";
 export * as stylistic from "./stylistic.js";