+++ test/bootstrap.cjs
@@ -1,3 +1,0 @@
-const must = require("must/register");
-// @ts-ignore
-global.must = must;