+++ CHANGELOG.md
@@ -7,6 +7,11 @@ Changes in rules are never considered breaking.
 ## [UNRELEASED]
 (nothing yet)
 
+## [9.37.2] - 2025-10-10
+### Changed
+- upgraded `object-shorthand` to always expect shorthand
+- always put `react` first when sorting imports
+
 ## [9.37.1] - 2025-10-09
 ### Removed
 - disabled `no-useless-return` (will enable when disabling auto-fixing becomes a thing)